Refactor: Change Layout 1st
