+++ src/data/network/APIEndpoints.js
@@ -1,0 +1,7 @@
+import {getIssue} from "./APIs/Issue.js";
+import {getLabel} from "./APIs/Label.js";
+
+export {
+  getIssue,
+  getLabel
+}
