+++ src/data/network/APIs/Issue.js
@@ -1,0 +1,4 @@
+import {getOptions, customFetch} from "../../repositories/utils/MyFetch.js";
+import $K from "../../../Constants.js"
+
+export const getIssue = customFetch(getOptions)($K.EndPoints.Issue.Get)
