+++ src/data/network/APIs/Label.js
@@ -1,0 +1,4 @@
+import {getOptions, customFetch} from "../../repositories/utils/MyFetch.js";
+import $K from "../../../Constants.js"
+
+export const getLabel = customFetch(getOptions)($K.EndPoints.Label.Get)
