+++ src/main.js
@@ -1,5 +1,5 @@
 import router, {app} from "./util/Router";
-import attachNavigationEvent from "./util/Navigator"
+import attachNavigationEvent from "./presentation/utils/Navigator.js"
 
 // MARK: init
 (() => {
