+++ src/presentation/utils/Navigator.js
@@ -1,4 +1,4 @@
-import router from "./Router";
+import router from "../../util/Router.js";
 
 export default function attachNavigationEvent() {
   const nav = document.getElementsByTagName('nav')[0]
