--- src/util/APIs/Issue.js
@@ -1,4 +1,0 @@
-import {getOptions, customFetch} from "../../data/repositories/utils/MyFetch.js";
-import $K from "../../Constants"
-
-export const getIssue = customFetch(getOptions)($K.EndPoints.Issue.Get)
