--- src/util/APIs/Label.js
@@ -1,4 +1,0 @@
-import {getOptions, customFetch} from "../../data/repositories/utils/MyFetch.js";
-import $K from "../../Constants"
-
-export const getLabel = customFetch(getOptions)($K.EndPoints.Label.Get)
