+++ src/view/LabelView.js
@@ -1,4 +1,4 @@
-import {getLabel} from "../util/APIs/Label";
+import {getLabel} from "../data/network/APIEndpoints";
 import {clearLabelBeforeRender, renderLabelList} from "../components/label/LabelItem";
 import {pipe} from "../util/FP";
